Non-negative Matrix Factorization for Binary Data

Non-negative Matrix Factorization for Binary Data

Usage

nmfbin(
  V,
  k,
  divergence_type = "kl",
  init_method = "random",
  max.iter = 100,
  tol = 1e-05,
  learning_rate = 0.01,
  compute_marginal_divergences = FALSE
)

Arguments

V

A binary matrix (with entries 0 or 1) to factorize.

k

Number of components.

divergence_type

Method for divergence calculation. Choices are "kl" for Kullback-Leibler divergence and "crossentropy" for binary cross-entropy.

init_method

Method for matrix factor initialization. Choices are "svd" for Singular Value Decomposition and "random" for random values.

max.iter

Maximum number of iterations for the factor update algorithm.

tol

Tolerance for convergence in the factor update algorithm.

learning_rate

Learning rate for gradient descent (only used with divergence_type = "crossentropy").

compute_marginal_divergences

Logical indicating whether to compute the marginal divergences.

Value

A list containing the factor matrices W and H, the final divergence, and if compute_marginal_divergences=TRUE, a vector of marginal divergences.

Examples

# Create a small binary matrix
V <- matrix(c(1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1), nrow = 3)

# Perform binary NMF with 2 components using KL divergence and random initialization
result <- nmfbin(V, k = 2, divergence_type = "kl", init_method = "random")

# Print factor matrices
print(result$W)
#>              [,1]         [,2]
#> [1,] 1.005274e+00 3.992284e-31
#> [2,] 3.204819e-09 1.382856e+00
#> [3,] 6.701828e-01 4.959178e-14
print(result$H)
#>           [,1]      [,2]      [,3]      [,4]      [,5]      [,6]
#> [1,] 0.6474869 0.6474869 0.0000000 0.0000000 0.6474869 1.2949737
#> [2,] 0.0000000 0.7844557 0.7844557 0.7844557 0.7844557 0.7844557

# Check the final divergence
print(result$final_divergence)
#> [1] 9.789299